perf(ui/sidebar): optimize setup process and i18n lazy load
