+++ ui/artalk-sidebar/src/App.vue
@@ -1,86 +1,10 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia'
-import type Artalk from 'artalk'
 import { useNavStore } from './stores/nav'
-import { useUserStore } from './stores/user'
-import { getArtalk, bootParams } from './global'
+import { bootParams } from './global'
 
 const nav = useNavStore()
-const user = useUserStore()
-const route = useRoute()
-const router = useRouter()
 const { scrollableArea } = storeToRefs(nav)
-
-const artalkLoaded = ref(false)
-
-onBeforeMount(() => {
-  const artalk = getArtalk()
-  if (!artalk) {
-    throw new Error('Artalk instance not initialized')
-  }
-
-  artalk.on('mounted', () => {
-    if (artalkLoaded.value) return
-    artalkLoaded.value = true
-
-    syncArtalk(artalk)
-  })
-})
-
-function syncArtalk(artalk: Artalk) {
-  // access from open sidebar or directly by url
-  if (bootParams.user?.email) {
-    // sync user from sidebar to artalk
-    artalk.ctx.get('user').update(bootParams.user)
-  } else {
-    // sync user from artalk to sidebar
-    try {
-      useUserStore().sync()
-    } catch {
-      nextTick(() => {
-        router.replace('/login')
-      })
-      return
-    }
-  }
-
-  // 验证登录身份有效性
-  const artalkUser = artalk.ctx.get('user')
-  const artalkUserData = artalkUser.getData()
-
-  const logout = () => {
-    user.logout()
-    nextTick(() => {
-      router.replace('/login')
-    })
-  }
-
-  // Remove login failed dialog if sidebar
-  artalk.ctx.getApiHandlers().remove('need_login')
-  artalk.ctx.getApiHandlers().add('need_login', async () => {
-    logout()
-    throw new Error('Need login')
-  })
-
-  // Check user status
-  artalk.ctx
-    .getApi()
-    .user.getUserStatus({
-      email: artalkUserData.email,
-      name: artalkUserData.name,
-    })
-    .then((res) => {
-      if (res.data.is_admin && !res.data.is_login) {
-        logout()
-      } else {
-        // 将全部通知标记为已读
-        artalk.ctx.getApi().notifies.markAllNotifyRead({
-          email: artalkUserData.email,
-          name: artalkUserData.name,
-        })
-      }
-    })
-}
 
 const darkMode = ref(bootParams.darkMode)
 
@@ -94,11 +18,7 @@ const darkMode = ref(bootParams.darkMode)
 </script>
 
 <template>
-  <div
-    v-if="artalkLoaded"
-    class="app-wrap artalk atk-sidebar"
-    :class="{ 'atk-dark-mode': darkMode }"
-  >
+  <div class="app-wrap artalk atk-sidebar" :class="{ 'atk-dark-mode': darkMode }">
     <AppHeader />
     <AppNavigation />
 
@@ -137,7 +57,7 @@ $sidebarWidth: 280px;
     }
   }
 
-  // 分页条占位
+  // The placeholder area for the pagination bar
   :deep(.atk-pagination-wrap) {
     z-index: 200;
     position: fixed;
