+++ ui/artalk-sidebar/src/i18n-en.ts
@@ -1,0 +1,99 @@
+export const en = {
+  ctrlCenter: 'Admin',
+  msgCenter: 'Messages',
+  noContent: 'No Content',
+  searchHint: 'Search by keywords...',
+  allSites: 'All Sites',
+  siteManage: 'Site Management',
+  comment: 'Comment',
+  page: 'Page',
+  user: 'User',
+  site: 'Site',
+  transfer: 'Transfer',
+  settings: 'Settings',
+  all: 'All',
+  pending: 'Pending',
+  personal: 'Personal',
+  mentions: 'Mentions',
+  mine: 'Mine',
+  admin: 'Admin',
+  create: 'Create',
+  import: 'Import',
+  export: 'Export',
+  settingSaved: 'Setting saved',
+  settingSaveFailed: 'Setting save failed',
+  settingNotice: 'Note: Some config options may require a manual reboot to take effect.',
+  apply: 'Apply',
+  updateComplete: 'Update complete',
+  updateReady: 'Ready to update...',
+  opFailed: 'Operation failed',
+  updateTitle: 'Fetch Title',
+  uploading: 'Uploading',
+  cancel: 'Cancel',
+  back: 'Back',
+  cacheClear: 'Cache Clear',
+  cacheWarm: 'Cache Warm',
+  editTitle: 'Edit Title',
+  switchKey: 'Switch Key',
+  commentAllowAll: 'Anyone Comment',
+  commentOnlyAdmin: 'Admin Comment Only',
+  config: 'Config',
+  envVarControlHint: 'Referenced by the environment variable {key}',
+  userAdminHint: 'Admin user',
+  userInConfHint: 'This user is defined in config file',
+  edit: 'Edit',
+  delete: 'Delete',
+  siteCount: 'Total {count} Sites',
+  createSite: 'Create Site',
+  siteName: 'Site Name',
+  siteUrls: 'Site URLs',
+  multiSepHint: 'multiple separated by commas',
+  add: 'Add',
+  rename: 'Rename',
+  inputHint: 'Input text...',
+  userCreate: 'Create User',
+  userEdit: 'Edit User',
+  userInConfCannotEditHint:
+    'Cannot edit user in config file, please modify the config file manually',
+  userDeleteConfirm:
+    'This operation will delete all comments of user: "{name}" email: "{email}", including the reply comments under his comments. Continue?',
+  userDeleteManuallyHint:
+    'User has been deleted from the database, please manually edit the config file and delete the user',
+  pageDeleteConfirm:
+    'This operation will delete the page: "{title}" and all data under it. Continue?',
+  siteDeleteConfirm:
+    'This operation will delete the site: "{name}" and all data under it. Continue?',
+  siteNameInputHint: 'Please enter the site name',
+  comments: 'Comments',
+  last: 'Last',
+  show: 'Show',
+  username: 'Username',
+  email: 'Email',
+  link: 'Link',
+  badgeText: 'Badge Text',
+  badgeColor: 'Badge Color',
+  role: 'Role',
+  normal: 'Normal',
+  password: 'Password',
+  passwordEmptyHint: 'leave blank not change your password',
+  emailNotify: 'Email notification',
+  enabled: 'Enabled',
+  disabled: 'Disabled',
+  save: 'Save',
+  dataFile: 'Data File',
+  artransfer: 'Artransfer',
+  targetSiteName: 'Target Site Name',
+  targetSiteURL: 'Target Site URL',
+  payload: 'Payload',
+  optional: 'Optional',
+  uploadReadyToImport: 'File uploaded and is ready for import',
+  artransferToolHint: 'Use the {link} to convert data to Artrans format.',
+  moreDetails: 'More details',
+  loginFailure: 'Login failure',
+  login: 'Login',
+  logout: 'Logout',
+  logoutConfirm: 'Are you sure you want to log out?',
+  loginSelectHint: 'Please select the account you wish to log into:',
+}
+
+export default en
