--- ui/artalk-sidebar/src/i18n/messages.ts
@@ -1,257 +1,0 @@
-const en = {
-  ctrlCenter: 'Admin',
-  msgCenter: 'Messages',
-  noContent: 'No Content',
-  searchHint: 'Search by keywords...',
-  allSites: 'All Sites',
-  siteManage: 'Site Management',
-  comment: 'Comment',
-  page: 'Page',
-  user: 'User',
-  site: 'Site',
-  transfer: 'Transfer',
-  settings: 'Settings',
-  all: 'All',
-  pending: 'Pending',
-  personal: 'Personal',
-  mentions: 'Mentions',
-  mine: 'Mine',
-  admin: 'Admin',
-  create: 'Create',
-  import: 'Import',
-  export: 'Export',
-  settingSaved: 'Setting saved',
-  settingSaveFailed: 'Setting save failed',
-  settingNotice: 'Note: Some config options may require a manual reboot to take effect.',
-  apply: 'Apply',
-  updateComplete: 'Update complete',
-  updateReady: 'Ready to update...',
-  opFailed: 'Operation failed',
-  updateTitle: 'Fetch Title',
-  cacheClear: 'Cache Clear',
-  cacheWarm: 'Cache Warm',
-  editTitle: 'Edit Title',
-  switchKey: 'Switch Key',
-  commentAllowAll: 'Anyone Comment',
-  commentOnlyAdmin: 'Admin Comment Only',
-  config: 'Config',
-  envVarControlHint: 'Referenced by the environment variable {key}',
-  userAdminHint: 'Admin user',
-  userInConfHint: 'This user is defined in config file',
-  edit: 'Edit',
-  delete: 'Delete',
-  siteCount: 'Total {count} Sites',
-  createSite: 'Create Site',
-  siteName: 'Site Name',
-  siteUrls: 'Site URLs',
-  multiSepHint: 'multiple separated by commas',
-  add: 'Add',
-  rename: 'Rename',
-  inputHint: 'Input text...',
-  userCreate: 'Create User',
-  userEdit: 'Edit User',
-  comments: 'Comments',
-  last: 'Last',
-  show: 'Show',
-  username: 'Username',
-  email: 'Email',
-  link: 'Link',
-  badgeText: 'Badge Text',
-  badgeColor: 'Badge Color',
-  role: 'Role',
-  normal: 'Normal',
-  password: 'Password',
-  passwordEmptyHint: 'leave blank not change your password',
-  emailNotify: 'Email notification',
-  enabled: 'Enabled',
-  disabled: 'Disabled',
-  save: 'Save',
-  dataFile: 'Data File',
-  artransfer: 'Artransfer',
-  targetSiteName: 'Target Site Name',
-  targetSiteURL: 'Target Site URL',
-  payload: 'Payload',
-  optional: 'Optional',
-  uploadReadyToImport: 'File uploaded and is ready for import',
-  artransferToolHint: 'Use the {link} to convert data to Artrans format.',
-  moreDetails: 'More details',
-  loginFailure: 'Login failure',
-  login: 'Login',
-  logout: 'Logout',
-  logoutConfirm: 'Are you sure you want to log out?',
-  loginSelectHint: 'Please select the account you wish to log into:',
-}
-
-const zhCN: typeof en = {
-  ctrlCenter: '控制中心',
-  msgCenter: '通知中心',
-  noContent: '无内容',
-  searchHint: '搜索关键字...',
-  allSites: '所有站点',
-  siteManage: '站点管理',
-  comment: '评论',
-  page: '页面',
-  user: '用户',
-  site: '站点',
-  transfer: '迁移',
-  settings: '设置',
-  all: '全部',
-  pending: '待审',
-  personal: '个人',
-  mentions: '提及',
-  mine: '我的',
-  admin: '管理员',
-  create: '新增',
-  import: '导入',
-  export: '导出',
-  settingSaved: '配置已保存',
-  settingSaveFailed: '配置保存失败',
-  settingNotice: '注：某些配置项可能需手动重启才能生效',
-  apply: '应用',
-  updateComplete: '更新完毕',
-  updateReady: '开始更新...',
-  opFailed: '操作失败',
-  updateTitle: '更新标题',
-  cacheClear: '缓存清除',
-  cacheWarm: '缓存预热',
-  editTitle: '标题修改',
-  switchKey: 'KEY 变更',
-  commentAllowAll: '所有人可评',
-  commentOnlyAdmin: '仅管理员可评',
-  config: '配置文件',
-  envVarControlHint: '由环境变量 {key} 控制',
-  userAdminHint: '该用户具有管理员权限',
-  userInConfHint: '该用户存在于配置文件中',
-  edit: '编辑',
-  delete: '删除',
-  siteCount: '共 {count} 个站点',
-  createSite: '新增站点',
-  siteName: '站点名称',
-  siteUrls: '站点 URLs',
-  multiSepHint: '多个用逗号隔开',
-  add: '新增',
-  rename: '重命名',
-  inputHint: '输入内容...',
-  userCreate: '用户创建',
-  userEdit: '用户编辑',
-  comments: '评论',
-  last: '近期',
-  show: '展开',
-  username: '用户名',
-  email: '邮箱',
-  link: '链接',
-  badgeText: '徽章文字',
-  badgeColor: '徽章颜色',
-  role: '身份角色',
-  normal: '普通',
-  password: '密码',
-  passwordEmptyHint: '留空不修改密码',
-  emailNotify: '邮件通知',
-  enabled: '开启',
-  disabled: '关闭',
-  save: '保存',
-  dataFile: '数据文件',
-  artransfer: '转换工具',
-  targetSiteName: '目标站点名',
-  targetSiteURL: '目标站点 URL',
-  payload: '启动参数',
-  optional: '可选',
-  uploadReadyToImport: '文件已成功上传，可以开始导入',
-  artransferToolHint: '使用 {link} 将评论数据转为 Artrans 格式',
-  moreDetails: '查看详情',
-  loginFailure: '登录发生异常',
-  login: '登录',
-  logout: '退出登录',
-  logoutConfirm: '确定要退出登录吗？',
-  loginSelectHint: '请选择您要登录的账户：',
-}
-
-const zhTW: typeof en = {
-  ctrlCenter: '控制中心',
-  msgCenter: '訊息中心',
-  noContent: '無內容',
-  searchHint: '關鍵字搜尋...',
-  allSites: '所有網站',
-  siteManage: '網站管理',
-  comment: '評論',
-  page: '頁面',
-  user: '用戶',
-  site: '網站',
-  transfer: '轉移',
-  settings: '設置',
-  all: '全部',
-  pending: '待審',
-  personal: '個人',
-  mentions: '提及',
-  mine: '我的',
-  admin: '管理員',
-  create: '建立',
-  import: '匯入',
-  export: '匯出',
-  settingSaved: '設定已儲存',
-  settingSaveFailed: '設定儲存失敗',
-  settingNotice: '注：某些設定選項可能需要手動重啟才能生效',
-  apply: '套用',
-  updateComplete: '更新完畢',
-  updateReady: '準備更新...',
-  opFailed: '操作失敗',
-  updateTitle: '擷取標題',
-  cacheClear: '清除快取',
-  cacheWarm: '預熱快取',
-  editTitle: '編輯標題',
-  switchKey: 'KEY 變更',
-  commentAllowAll: '允許任何人評論',
-  commentOnlyAdmin: '僅允許管理員評論',
-  config: '配置文件',
-  envVarControlHint: '由環境變數 {key} 參照',
-  userAdminHint: '該用戶具有管理員權限',
-  userInConfHint: '該用戶存在於配置文件中',
-  edit: '編輯',
-  delete: '刪除',
-  siteCount: '共 {count} 個網站',
-  createSite: '建立網站',
-  siteName: '網站名稱',
-  siteUrls: '網站 URLs',
-  multiSepHint: '使用逗號分隔多個',
-  add: '新增',
-  rename: '重命名',
-  inputHint: '輸入文字...',
-  userCreate: '建立用戶',
-  userEdit: '用戶編輯',
-  comments: '評論',
-  last: '最後',
-  show: '展開',
-  username: '用戶名',
-  email: '郵箱',
-  link: '連結',
-  badgeText: '徽章文字',
-  badgeColor: '徽章顏色',
-  role: '身份角色',
-  normal: '一般',
-  password: '密碼',
-  passwordEmptyHint: '留空表示不變更密碼',
-  emailNotify: '郵件通知',
-  enabled: '啟用',
-  disabled: '停用',
-  save: '儲存',
-  dataFile: '資料檔案',
-  artransfer: '轉移工具',
-  targetSiteName: '目標網站名稱',
-  targetSiteURL: '目標網站 URL',
-  payload: '有效載荷',
-  optional: '選填',
-  uploadReadyToImport: '檔案已上傳並準備匯入',
-  artransferToolHint: '使用 {link} 將評論數據轉換為 Artrans 格式。',
-  moreDetails: '查看詳情',
-  loginFailure: '登入失敗',
-  login: '登入',
-  logout: '登出',
-  logoutConfirm: '確定要登出嗎？',
-  loginSelectHint: '請選擇您要登入的帳號：',
-}
-
-export default {
-  en,
-  'zh-CN': zhCN,
-  'zh-TW': zhTW,
-}
