+++ ui/artalk-sidebar/src/lib/promise-polyfill.ts
@@ -1,0 +1,9 @@
+Promise.withResolvers ??= function <T>() {
+  let resolve: PromiseWithResolvers<T>['resolve']
+  let reject: PromiseWithResolvers<T>['reject']
+  const promise = new Promise<T>((res, rej) => {
+    resolve = res
+    reject = rej
+  })
+  return { promise, resolve: resolve!, reject: reject! }
+}
